requirements.txt
requests
beautifulsoup4
PyPDF2
google-generativeai
python-dotenv